bv_ans/requirements.txt
#Atualizada no dia 13/11/2025
pylib-sgen-base-genai-framework==0.17.0
google-generativeai
google-ai-generativelanguage
google-genai==1.47.0
google-adk==1.18.0
google-cloud-aiplatform==1.126.1
 pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
python-dotenv==1.2.1
python-dateutil==2.9.0.post0
flask==2.3.0
fastapi==0.120.2
python-dotenv==1.2.1
python-multipart==0.0.20
pandas==2.2.3
openpyxl==3.1.5